requirements.txt
opencv-python>=4.5.0
mediapipe>=0.8.0
numpy>=1.20.0
pygame>=2.0.0
river>=0.15.0
pyserial>=3.5 